+++ commands/doc-start.js
@@ -3,11 +3,11 @@ const Webpack = require("webpack");
 const WebpackDevServer = require("webpack-dev-server");
 const generateConfig = require("../includes/webpack.config");
 
-module.exports = function () {
+module.exports = function (cmd) {
   const paths = getPaths("doc");
   const config = generateConfig({ paths, hot: true });
   const compiler = Webpack(config);
-  const port = 8080;
+  const port = cmd.port || 8080;
   const host = "127.0.0.1";
   const devServerOptions = {
     host,
@@ -24,6 +24,6 @@ module.exports = function () {
   const server = new WebpackDevServer(compiler, devServerOptions);
 
   server.listen(port, host, () => {
-    console.log("Starting server on http://localhost:8080");
+    console.log(`Starting server on http://${host}:${port}`);
   });
 };